requirements.txt
PyYAML
toml
jinja2
semver
